requirements.txt
discord.py
python-dotenv
pynacl
faster-whisper
discord-ext-voice-recv